backend/requirements.txt
# -----------------------------
# 🌐 FastAPI and Server
# -----------------------------
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.3.4

# -----------------------------
# 💾 Database (MongoDB)
# -----------------------------
motor==3.6.0
pymongo >= 4.9, < 4.10


# -----------------------------
# 🧠 AI / LLM Integrations
# -----------------------------
openai==1.47.0

# -----------------------------
# 🌍 HTTP Client
# -----------------------------
httpx==0.27.2

# -----------------------------
# ⚙️ Environment Variables
# -----------------------------
python-dotenv==1.0.1

# -----------------------------
# 🧩 Utilities
# -----------------------------
python-multipart==0.0.9

# -----------------------------
# 🔐 Auth & Security
# -----------------------------
python-jose[cryptography]==3.3.0
